--- Engine/src/boost_serialization_includes.h
@@ -1,4 +1,0 @@
-#pragma once
-#include <boost/archive/text_oarchive.hpp>
-#include <boost/archive/text_iarchive.hpp>
-#include <boost/serialization/vector.hpp>
