--- Engine/src/engine.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "achievements/achievement.h"
-#include "command/command.h"
-#include "item/item.h"
-#include "node/node.h"
-#include "player/inventory.h"
-#include "player/player.h"
-#include "progress/progress.h"
-#include "utils/game_utils.h"
-#include "utils/prompt.h"
-#include "text/text.h"
-#include "map/map.h"
-#include "menu/menu.h"
-#include "mission/mission.h"
-#include "mission/objective.h"
