+++ include/text.h
