+++ scripts/build
@@ -1,0 +1,12 @@
+#!/bin/zsh
+
+echo "Building project..."
+
+mkdir build
+cd build
+
+cmake ..
+
+make
+
+echo "Done"
