+++ scripts/clean
@@ -1,0 +1,7 @@
+#!/bin/zsh
+
+echo "Cleaning project..."
+
+rm -rf build
+
+echo "Done"
