+++ src/node/node.cpp
